+++ crates/prices/src/boundary.rs
@@ -1,11 +1,12 @@
 use {crate::core, futures::FutureExt};
 
+/// An estimator which simply delegates to the legacy code.
 pub struct Estimator(Box<dyn shared::price_estimation::PriceEstimating>);
 
 impl core::Estimator for Estimator {
     fn estimate(
         &self,
-        query: core::Query,
+        swap: core::Swap,
         deadline: core::Deadline,
     ) -> futures::future::BoxFuture<'_, Result<core::estimator::Estimate, core::estimator::Error>>
     {
@@ -15,9 +16,9 @@ impl core::Estimator for Estimator {
                 shared::price_estimation::single_estimate(
                     self.0.as_ref(),
                     &shared::price_estimation::Query {
-                        sell_token: query.from.into(),
-                        buy_token: query.to.into(),
-                        in_amount: query.amount.into(),
+                        sell_token: swap.from.into(),
+                        buy_token: swap.to.into(),
+                        in_amount: swap.amount.into(),
                         kind: model::order::OrderKind::Sell,
                         verification: None,
                     },
