+++ crates/prices/src/core/estimator.rs
@@ -1,26 +1,21 @@
 use {
-    super::{eth, Deadline, Query, ToAmount},
+    super::{eth, swap, Deadline, Swap},
     futures::future::BoxFuture,
 };
 
-/// A price estimator. Given a [`Query`] (which specifies how much of the
-/// [`crate::FromToken`] should be converted into the [`crate::ToToken`]), the
-/// estimator returns an [`Estimate`] of how much of the [`crate::ToToken`] will
-/// be received, and how much will be paid in gas fees.
+/// An external token price estimator. Given a [`Swap`] (which specifies how
+/// much of the [`crate::FromToken`] should be converted into the
+/// [`crate::ToToken`]), the estimator returns an [`Estimate`] of how much of
+/// the [`crate::ToToken`] will be received, and how much will be paid in gas
+/// fees.
 pub trait Estimator {
-    fn estimate(
-        &self,
-        query: Query,
-        // The deadline is intentionally not part of the query, to make it more difficult for
-        // estimator implementations to accidentally forget to use it.
-        deadline: Deadline,
-    ) -> BoxFuture<'_, Result<Estimate, Error>>;
+    fn estimate(&self, swap: Swap, deadline: Deadline) -> BoxFuture<'_, Result<Estimate, Error>>;
 }
 
 /// An estimate returned by an [`Estimator`].
 #[derive(Debug, Clone, Copy)]
 pub struct Estimate {
-    pub to: ToAmount,
+    pub to: swap::ToAmount,
     /// The gas cost of the swap.
     pub gas: eth::Gas,
 }
@@ -31,10 +26,7 @@ pub struct Estimate {
 pub struct Error(Box<dyn std::error::Error>);
 
 impl Error {
-    pub fn new<E>(err: E) -> Self
-    where
-        E: std::error::Error + 'static,
-    {
+    pub fn new(err: impl std::error::Error + 'static) -> Self {
         Self(Box::new(err))
     }
 }
